Separate copied/modified logic for maintenance and readability
